+++ CHANGELOG.md
@@ -32,6 +32,9 @@ Line wrap the file at 100 chars.                                              Th
 - Send an ICMP reject message or TCP reset packet when blocking outgoing packets to prevent
   timeouts.
 
+#### macOS
+- Use `SCNetworkReachability` to help determine connectivity of host.
+
 #### Android
 - Show the remaining account time in the Settings screen in days if it's less than 3 months.
 - Prevent commands to connect or disconnect to be sent when the device is locked.